+++ module/powerSheet.mjs
@@ -1,29 +1,31 @@
-import { ppText } from './utils.mjs';
+import {ppText} from "./utils.mjs";
 
 export default class PowerSheet extends dnd5e.applications.item.ItemSheet5e {
   get template() {
-    return `modules/prime-psionics/templates/power-sheet.hbs`;
+    return "modules/prime-psionics/templates/power-sheet.hbs";
   }
 
   async getData(options = {}) {
     const context = await super.getData(options);
     context.psionics = CONFIG.PSIONICS;
+
     context.powerComponents = {
       ...CONFIG.PSIONICS.powerComponents,
-      ...CONFIG.DND5E.spellTags,
+      concentration: CONFIG.DND5E.itemProperties.concentration,
+      ritual: CONFIG.DND5E.itemProperties.ritual
     };
-    if (context.system.actionType === 'msak')
+    if (context.system.actionType === "msak")
       context.itemProperties[0] = game.i18n.localize(
-        'PrimePsionics.ActionMPAK'
+        "PrimePsionics.ActionMPAK"
       );
-    if (context.system.actionType === 'rsak')
+    if (context.system.actionType === "rsak")
       context.itemProperties[0] = game.i18n.localize(
-        'PrimePsionics.ActionRPAK'
+        "PrimePsionics.ActionRPAK"
       );
 
     const consume =
-      context.system.consume.type === 'flags'
-        ? { pp: game.i18n.localize('PrimePsionics.PP') }
+      context.system.consume.type === "flags"
+        ? {pp: game.i18n.localize("PrimePsionics.PP")}
         : {};
 
     context.powerScalingModes = CONFIG.PSIONICS.powerScalingModes;
@@ -38,7 +40,7 @@ export default class PowerSheet extends dnd5e.applications.item.ItemSheet5e {
     } else delete context.system.labels.pp;
     foundry.utils.mergeObject(context, {
       labels: context.system.labels,
-      abilityConsumptionTargets: consume,
+      abilityConsumptionTargets: consume
     });
 
     return context;