+++ rollup.config.mjs
@@ -1,7 +1,7 @@
 export default {
-  input: './prime-psionics.mjs',
+  input: "./prime-psionics.mjs",
   output: {
-    file: './prime-psionics-compiled.mjs',
-    format: 'esm',
-  },
+    file: "./prime-psionics-compiled.mjs",
+    format: "esm"
+  }
 };